requirements.txt
fastapi
uvicorn
python-dotenv
openai
langchain
langchain-community